Refactor type index into a separate module
